OrlandoDiaz_Ex5
I worked on classes and objects for our class monster and a new one I
made to test out making the classes and objects more independently. The
class one has the aliens and mine has a snowman with a screen with Bill
Cipher (again :P) on it. I cloned the aliens 5 times and the
snowman/Bill combo 4 times. I hope you like them!